Adds basic handling for friction
Modified hameq to also include a basic handler for electronic friction.

This uses the 3rd PIMPAR parameter, if there are not 3 PIMPAR parameters, it is considered as 0, and is not used.

The forumla for energy loss used is:
dE = -v*2 Q dt, where Q is the PIMPAR parameter.
The momentum loss is then:
dp_i = -v_i * sqrt(2*m*Q*dt)